Move config logic into its folder
